Move shared fields and methods from V1/V2 into GsymCreator/GsymReader base classes
Creator: moved all fields and ~25 identical methods to GsymCreator base.
Only encode(), calculateHeaderAndTableSize(), and loadCallSitesFromYAML()
remain as virtual overrides. Added createNew() virtual factory for
createSegment() support.

Reader: moved shared fields (MemBuffer, Endian, AddrOffsets, AddrInfoOffsets,
Files, StrTab) and cached header values (BaseAddress, NumAddresses, AddrOffSize)
to GsymReader base. Moved ~20 identical methods including getAddress(),
getFunctionInfoDataForAddress(), lookup(), lookupAll(), and all dump overloads
(except dump(OS) which prints the version-specific header). Only parse() and
getHeader() remain in subclasses.

User prompt:
"Can you see if the v1/v2 classes share the same members (fields and methods),
and move them into the base classes?"
