+++ llvm/include/llvm/DebugInfo/GSYM/GsymReader.h
@@ -30,85 +30,119 @@ class raw_ostream;
 
 namespace gsym {
 
-/// GsymReader is an abstract interface for reading GSYM data.
+/// GsymReader is the base class for reading GSYM data.
 ///
-/// This interface provides the methods needed by FunctionInfo::lookup and
-/// InlineInfo::lookup to resolve strings and files during symbolication.
-/// Both GsymReaderV1 and GsymReaderV2 implement this interface.
+/// This class contains all shared state and logic for V1 and V2 readers.
+/// Subclasses implement version-specific parsing (parse()) and header access.
 class GsymReader {
+protected:
+  std::unique_ptr<MemoryBuffer> MemBuffer;
+  llvm::endianness Endian;
+  ArrayRef<uint8_t> AddrOffsets;
+  ArrayRef<uint32_t> AddrInfoOffsets;
+  ArrayRef<FileEntry> Files;
+  StringTable StrTab;
+
+  // Cached header values, populated by subclass parse().
+  uint64_t CachedBaseAddress = 0;
+  uint32_t CachedNumAddresses = 0;
+  uint8_t CachedAddrOffSize = 0;
+
+  LLVM_ABI GsymReader(std::unique_ptr<MemoryBuffer> Buffer);
+
+  template <class T> ArrayRef<T>
+  getAddrOffsets() const {
+    return ArrayRef<T>(reinterpret_cast<const T *>(AddrOffsets.data()),
+                       AddrOffsets.size()/sizeof(T));
+  }
+
+  template <class T>
+  std::optional<uint64_t> addressForIndex(size_t Index) const {
+    ArrayRef<T> AIO = getAddrOffsets<T>();
+    if (Index < AIO.size())
+      return AIO[Index] + CachedBaseAddress;
+    return std::nullopt;
+  }
+
+  template <class T>
+  std::optional<uint64_t>
+  getAddressOffsetIndex(const uint64_t AddrOffset) const {
+    ArrayRef<T> AIO = getAddrOffsets<T>();
+    const auto Begin = AIO.begin();
+    const auto End = AIO.end();
+    auto Iter = std::lower_bound(Begin, End, AddrOffset);
+    if (Iter == Begin && AddrOffset < *Begin)
+      return std::nullopt;
+    if (Iter == End || AddrOffset < *Iter)
+      --Iter;
+
+    while (Iter != Begin) {
+      auto Prev = Iter - 1;
+      if (*Prev == *Iter)
+        Iter = Prev;
+      else
+        break;
+    }
+
+    return std::distance(Begin, Iter);
+  }
+
+  LLVM_ABI Expected<uint64_t> getAddressIndex(const uint64_t Addr) const;
+  LLVM_ABI std::optional<uint64_t> getAddressInfoOffset(size_t Index) const;
+  LLVM_ABI llvm::Expected<llvm::DataExtractor>
+  getFunctionInfoDataForAddress(uint64_t Addr, uint64_t &FuncStartAddr) const;
+  LLVM_ABI llvm::Expected<llvm::DataExtractor>
+  getFunctionInfoDataAtIndex(uint64_t AddrIdx, uint64_t &FuncStartAddr) const;
+
 public:
+  LLVM_ABI GsymReader(GsymReader &&RHS);
   virtual ~GsymReader() = default;
 
   /// Open a GSYM file, auto-detecting the format version.
-  ///
-  /// \param Path The file path of the GSYM file to read.
-  /// \returns An expected unique_ptr to a GsymReader or an error.
   LLVM_ABI static llvm::Expected<std::unique_ptr<GsymReader>>
   openFile(StringRef Path);
 
   /// Construct a GsymReader from a buffer, auto-detecting the format version.
-  ///
-  /// \param Bytes A set of bytes that will be copied and owned by the
-  /// returned object on success.
-  /// \returns An expected unique_ptr to a GsymReader or an error.
   LLVM_ABI static llvm::Expected<std::unique_ptr<GsymReader>>
   copyBuffer(StringRef Bytes);
 
-  /// Get a string from the string table.
-  virtual StringRef getString(uint32_t Offset) const = 0;
+  StringRef getString(uint32_t Offset) const { return StrTab[Offset]; }
 
-  /// Get a file entry for the supplied file index.
-  virtual std::optional<FileEntry> getFile(uint32_t Index) const = 0;
+  std::optional<FileEntry> getFile(uint32_t Index) const {
+    if (Index < Files.size())
+      return Files[Index];
+    return std::nullopt;
+  }
 
-  /// Get the full function info for an address.
-  virtual llvm::Expected<FunctionInfo> getFunctionInfo(uint64_t Addr) const = 0;
+  uint32_t getNumAddresses() const { return CachedNumAddresses; }
 
-  /// Get the full function info given an address index.
-  virtual llvm::Expected<FunctionInfo>
-  getFunctionInfoAtIndex(uint64_t AddrIdx) const = 0;
+  LLVM_ABI llvm::Expected<FunctionInfo> getFunctionInfo(uint64_t Addr) const;
+  LLVM_ABI llvm::Expected<FunctionInfo>
+  getFunctionInfoAtIndex(uint64_t AddrIdx) const;
 
-  /// Lookup an address in the GSYM.
-  virtual llvm::Expected<LookupResult>
+  LLVM_ABI llvm::Expected<LookupResult>
   lookup(uint64_t Addr,
-         std::optional<DataExtractor> *MergedFuncsData = nullptr) const = 0;
+         std::optional<DataExtractor> *MergedFuncsData = nullptr) const;
 
-  /// Lookup all merged functions for a given address.
-  virtual llvm::Expected<std::vector<LookupResult>>
-  lookupAll(uint64_t Addr) const = 0;
+  LLVM_ABI llvm::Expected<std::vector<LookupResult>>
+  lookupAll(uint64_t Addr) const;
 
-  /// Get the number of addresses in this GSYM file.
-  virtual uint32_t getNumAddresses() const = 0;
+  LLVM_ABI std::optional<uint64_t> getAddress(size_t Index) const;
 
-  /// Gets an address from the address table.
-  virtual std::optional<uint64_t> getAddress(size_t Index) const = 0;
-
-  /// Dump the entire GSYM data contained in this object.
+  /// Dump the entire GSYM data. Version-specific (header format differs).
   virtual void dump(raw_ostream &OS) = 0;
 
-  /// Dump a FunctionInfo object.
-  virtual void dump(raw_ostream &OS, const FunctionInfo &FI,
-                    uint32_t Indent = 0) = 0;
-
-  /// Dump a MergedFunctionsInfo object.
-  virtual void dump(raw_ostream &OS, const MergedFunctionsInfo &MFI) = 0;
-
-  /// Dump a CallSiteInfo object.
-  virtual void dump(raw_ostream &OS, const CallSiteInfo &CSI) = 0;
-
-  /// Dump a CallSiteInfoCollection object.
-  virtual void dump(raw_ostream &OS, const CallSiteInfoCollection &CSIC,
-                    uint32_t Indent = 0) = 0;
-
-  /// Dump a LineTable object.
-  virtual void dump(raw_ostream &OS, const LineTable &LT,
-                    uint32_t Indent = 0) = 0;
-
-  /// Dump a InlineInfo object.
-  virtual void dump(raw_ostream &OS, const InlineInfo &II,
-                    uint32_t Indent = 0) = 0;
-
-  /// Dump a FileEntry object.
-  virtual void dump(raw_ostream &OS, std::optional<FileEntry> FE) = 0;
+  LLVM_ABI void dump(raw_ostream &OS, const FunctionInfo &FI,
+                     uint32_t Indent = 0);
+  LLVM_ABI void dump(raw_ostream &OS, const MergedFunctionsInfo &MFI);
+  LLVM_ABI void dump(raw_ostream &OS, const CallSiteInfo &CSI);
+  LLVM_ABI void dump(raw_ostream &OS, const CallSiteInfoCollection &CSIC,
+                     uint32_t Indent = 0);
+  LLVM_ABI void dump(raw_ostream &OS, const LineTable &LT,
+                     uint32_t Indent = 0);
+  LLVM_ABI void dump(raw_ostream &OS, const InlineInfo &II,
+                     uint32_t Indent = 0);
+  LLVM_ABI void dump(raw_ostream &OS, std::optional<FileEntry> FE);
 };
 
 } // namespace gsym
